AdapterTest and PresenterTest now separated
